agent/requirements.txt
openai==1.13.3
google-generativeai==0.4.1
requests==2.31.0
urllib3==2.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
pydantic==1.10.14

# Dev tooling
pytest==7.4.4
black==24.2.0
ruff==0.2.2
mypy==1.8.0
types-PyYAML==6.0.12.12
